+++ Tests/SentryTests/Integrations/WatchdogTerminations/SentryWatchdogTerminationTrackerTests.swift
@@ -27,6 +27,8 @@ class SentryWatchdogTerminationTrackerTests: NotificationCenterTestCase {
         let contextProcessor: SentryWatchdogTerminationContextProcessorWrapper
         let userProcessor: SentryWatchdogTerminationUserProcessorWrapper
         let tagsProcessor: SentryWatchdogTerminationTagsProcessorWrapper
+        let distProcessor: SentryWatchdogTerminationDistProcessorWrapper
+        let environmentProcessor: SentryWatchdogTerminationEnvironmentProcessorWrapper
 
         init() {
             SentryDependencyContainer.sharedInstance().sysctlWrapper = sysctl
@@ -51,6 +53,14 @@ class SentryWatchdogTerminationTrackerTests: NotificationCenterTestCase {
             tagsProcessor = SentryWatchdogTerminationTagsProcessorWrapper(
                 withDispatchQueueWrapper: backgroundQueueWrapper,
                 scopeTagsStore: SentryScopeTagsPersistentStore(fileManager: fileManager)
+            )
+            distProcessor = SentryWatchdogTerminationDistProcessorWrapper(
+                withDispatchQueueWrapper: backgroundQueueWrapper,
+                scopeDistStore: SentryScopeDistPersistentStore(fileManager: fileManager)
+            )
+            environmentProcessor = SentryWatchdogTerminationEnvironmentProcessorWrapper(
+                withDispatchQueueWrapper: backgroundQueueWrapper,
+                scopeEnvironmentStore: SentryScopeEnvironmentPersistentStore(fileManager: fileManager)
             )
 
             client = TestClient(options: options)
@@ -87,6 +97,12 @@ class SentryWatchdogTerminationTrackerTests: NotificationCenterTestCase {
             let scopeTagsPersistentStore = SentryScopeTagsPersistentStore(
                 fileManager: fileManager
             )
+            let scopeDistPersistentStore = SentryScopeDistPersistentStore(
+                fileManager: fileManager
+            )
+            let scopeEnvironmentPersistentStore = SentryScopeEnvironmentPersistentStore(
+                fileManager: fileManager
+            )
             return SentryWatchdogTerminationTracker(
                 options: options,
                 watchdogTerminationLogic: logic,
@@ -95,7 +111,9 @@ class SentryWatchdogTerminationTrackerTests: NotificationCenterTestCase {
                 fileManager: fileManager,
                 scopeContextStore: scopePersistentStore,
                 scopeUserStore: scopeUserPersistentStore,
-                scopeTagsStore: scopeTagsPersistentStore
+                scopeTagsStore: scopeTagsPersistentStore,
+                scopeDistStore: scopeDistPersistentStore,
+                scopeEnvironmentStore: scopeEnvironmentPersistentStore
             )
         }
     }
@@ -308,7 +326,9 @@ class SentryWatchdogTerminationTrackerTests: NotificationCenterTestCase {
             breadcrumbProcessor: fixture.breadcrumbProcessor,
             contextProcessor: fixture.contextProcessor,
             userProcessor: fixture.userProcessor,
-            tagsProcessor: fixture.tagsProcessor
+            tagsProcessor: fixture.tagsProcessor,
+            distProcessor: fixture.distProcessor,
+            environmentProcessor: fixture.environmentProcessor
         )
 
         for _ in 0..<3 {